requirements.txt
Flask==3.1.2
Jinja2==3.1.6
markdown2==2.5.4
pillow==12.1.0
psutil==7.2.2
python-dotenv==1.1.0
